Ajouter automatiquement le créateur dans les contacts
Dans SSA à la création d'un événement produit ajouter le créateur et sa
structure dans les contacts de la fiche (iso SV).
