+++ sv/view_mixins.py
@@ -1,7 +1,6 @@
 from django.core.exceptions import ValidationError
 from collections import defaultdict
 
-from core.models import Contact
 from sv.forms import (
     PrelevementForm,
 )
@@ -84,20 +83,6 @@ class WithPrelevementHandlingMixin:
                 for field, error in prelevement_form.errors.items():
                     error_msg += f"{field.title()}: {error.as_text()}\n"
                 raise ValidationError(error_msg)
-
-
-class WithAddUserContactsMixin:
-    """Mixin pour ajouter automatiquement l'utilisateur courant et sa structure comme contacts."""
-
-    def add_user_contacts(self, obj):
-        """Ajoute l'utilisateur courant et sa structure comme contacts de l'objet."""
-        agent = self.request.user.agent
-
-        agent_contact = Contact.objects.get(agent=agent)
-        obj.contacts.add(agent_contact)
-
-        if structure_contact := agent_contact.get_structure_contact():
-            obj.contacts.add(structure_contact)
 
 
 class WithPrelevementResultatsMixin:
